restructure: deps belong to gd
