+++ rgdps/api/gd/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
 
+from . import dependencies
 from . import responses
 from . import routes
