+++ rgdps/api/gd/routes/levels.py
@@ -8,7 +8,7 @@ from fastapi import Form
 from rgdps import logger
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.api.validators import Base64String
 from rgdps.api.validators import CommaSeparatedIntList
 from rgdps.api.validators import TextBoxString
