+++ rgdps/api/gd/routes/messages.py
@@ -5,7 +5,7 @@ from rgdps import logger
 from rgdps.api import commands
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.api.validators import Base64String
 from rgdps.api.validators import MessageContentString
 from rgdps.common import gd_obj
