+++ rgdps/api/gd/routes/rewards.py
@@ -6,7 +6,7 @@ from fastapi import Form
 from rgdps import logger
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.common import gd_obj
 from rgdps.constants.daily_chests import DailyChestView
 from rgdps.constants.errors import ServiceError
