+++ rgdps/api/gd/routes/save_data.py
@@ -8,7 +8,7 @@ from rgdps import logger
 from rgdps import settings
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.api.validators import GameSaveData
 from rgdps.constants.errors import ServiceError
 from rgdps.models.user import User
