+++ rgdps/api/gd/routes/user_comments.py
@@ -7,7 +7,7 @@ from rgdps import logger
 from rgdps.api import commands
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.api.validators import Base64String
 from rgdps.common import gd_obj
 from rgdps.constants.errors import ServiceError
