+++ rgdps/api/gd/routes/users.py
@@ -7,7 +7,7 @@ from pydantic import EmailStr
 from rgdps import logger
 from rgdps.api import responses
 from rgdps.api.context import HTTPContext
-from rgdps.api.dependencies import authenticate_dependency
+from rgdps.api.gd.dependencies import authenticate_dependency
 from rgdps.api.validators import SocialMediaString
 from rgdps.api.validators import TextBoxString
 from rgdps.common import gd_obj
